requirements.txt
Flask==2.0.1
gunicorn==20.1.0
Jinja2==3.1.2
matplotlib==3.7.1
numpy==1.24.2
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.3.0
pipenv==2023.3.20
PyGObject==3.42.1
requests==2.28.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
transformers==4.27.3
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3

